Remove PersonStorageService since it is only used from one place
